move ABI into separate module
